fix(checker): move boolean array display query to solver
