+++ crates/tsz-checker/src/tests/architecture_contract_tests_parts/part_05.rs
@@ -58,3 +58,32 @@ fn test_compiler_managed_name_predicate_has_domain_boundary() {
         "checker code must call query_boundaries::type_predicates::is_compiler_managed_type; violations: {violations:?}"
     );
 }
+
+/// The `RelationFailure` enum must live in `relation_types.rs` and provide
+/// structured variant coverage for the semantic families we're unifying.
+#[test]
+fn test_relation_failure_covers_semantic_families() {
+    let source = fs::read_to_string("src/query_boundaries/relation_types.rs")
+        .expect("failed to read query_boundaries/relation_types.rs");
+
+    // Core semantic families that must be represented
+    for variant in [
+        "MissingProperty",
+        "MissingProperties",
+        "ExcessProperty",
+        "IncompatiblePropertyValue",
+        "NoApplicableSignature",
+        "TupleArityMismatch",
+        "ReturnTypeMismatch",
+        "ParameterTypeMismatch",
+        "ParameterCountMismatch",
+        "PropertyModifierMismatch",
+        "WeakUnionViolation",
+        "TypeMismatch",
+    ] {
+        assert!(
+            source.contains(variant),
+            "RelationFailure must include the `{variant}` variant for semantic coverage"
+        );
+    }
+}
